ZeroTrust_SIEM/requirements.txt
flask
pymongo
customtkinter
tkintermapview
matplotlib
pandas
faker
pillow
scikit-learn
packaging